% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AT_Functions.R
\name{AlphaTestSRSFSP}
\alias{AlphaTestSRSFSP}
\title{Alpha Testing -
Single Return Horizon, Single Factor, Single Period}
\usage{
AlphaTestSRSFSP(data)
}
\arguments{
\item{data}{dataframe containing id columns, return column and factor value column}
}
\value{
None
}
\description{
This is the most baseline AT function.
}
\details{
This function returns for the universe and each quantile:
     number of observations \cr
     average return \cr
     average relative return for quantiles \cr
     security level hit rate of universe vs 0 \cr
     security level hit rate of universe vs universe average \cr
     information coefficient based on Z-Score of factor \cr
     information coefficient based on rank of factor \cr
     quantile spread between top and bottom quantile \cr
     regression alpha - universe only \cr
     regression beta - universe only
}
