% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Sort_Scoring.R
\name{ctq}
\alias{ctq}
\title{Convert Data to Quantiles}
\usage{
ctq(x, fftile)
}
\arguments{
\item{x}{vector to quantiled}

\item{fftile}{integer number of fractiles to use in splitting data}
}
\value{
retval
}
\description{
Function to calculate quantiles.
}
\details{
Used in Alpha Testing Functions
}
