% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Sort_Scoring.R
\name{ctz}
\alias{ctz}
\title{Calculate Z-Score with Winsorization}
\usage{
ctz(x, win.prob = c(0, 1))
}
\arguments{
\item{x}{a numeric vector to be winsorized and normalized.}

\item{win.prob}{numeric vector of probabilities with values in [0,1]
as used in quantile.}
}
\value{
A vector of the same length as the original data x containing the
winsorized and normalized data.
}
\description{
Function to calculate normalized value with windsorization.
}
\details{
Used in Alpha Testing Functions
}
