% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util_Sort_Scoring.R
\name{f_scoring}
\alias{f_scoring}
\title{Quantile & Z-Scoring}
\usage{
f_scoring(data, fname, fftile, winsor = c(0, 1))
}
\arguments{
\item{data}{dataframe containing column with data to be scored}

\item{fname}{character column name of factor}

\item{fftile}{integer number of fractiles to use in spliting data}

\item{winsor}{pair of numeric to bound data in windsorization}
}
\value{
data
}
\description{
Add Windsorized Z-Score and Quantile Score to Data
}
\details{
Used in Alpha Testing Functions
}
