% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{named_group_split}
\alias{named_group_split}
\title{Names Group Split}
\usage{
named_group_split(...)
}
\arguments{
\item{...}{dynamic dots}
}
\description{
Splits Data by Group into List and Names
}
\details{
Splits Data by Group into List and Names
}
